test(server): extract heartbeat run quiescence drain to a shared helper
Move drainHeartbeatRunsToQuiescence out of the low-trust red-team route
suite into server/src/__tests__/helpers/drain-heartbeat-runs.ts so the
sibling embedded-Postgres suites can reuse one drain primitive. The
low-trust suite now imports the helper and drops its inline copy.
